Fixing issues with PR
