+++ src/tests/testhelper.ts
@@ -1,4 +1,14 @@
-import { GameState, PlayerState, Spec } from "../framework/types";
+import { lookupCard } from "../data";
+import { performAction } from "../framework/actions";
+import {
+  CardID,
+  GameState,
+  InstanceID,
+  PlayerState,
+  Spec,
+} from "../framework/types";
+import { InstanceQuery, queryInstances } from "../framework/queries/common";
+
 
 export const P1 = "P1";
 export const P2 = "P2";
@@ -62,4 +72,36 @@ const generatePlayer = (
       },
     ],
   };
+};
+
+
+const findInstance = (
+  $: GameState,
+  query: InstanceQuery,
+): InstanceID => {
+  return queryInstances($, query)[0];
+};
+
+export const debugPlayCard = (
+  $: GameState,
+  cid: CardID,
+): void => {
+  const card = lookupCard(cid);
+  $.players[$.activePlayer].gold += card.cost;
+  $.players[$.activePlayer].hand.push(cid);
+  performAction($, {
+    type: "PLAY_CARD",
+    cardID: cid,
+    boost: false,
+  });
+};
+
+export const debugPlayUnit = (
+  $: GameState,
+  cid: CardID,
+): InstanceID => {
+  debugPlayCard($, cid);
+  return findInstance($, {
+    card: cid,
+  });
 };
